+++ ArcExplorer.Test/GetCleanedDirectoryPath.cs
@@ -9,43 +9,43 @@ namespace ArcExplorer.Test
         [TestMethod]
         public void NullPath()
         {
-            Assert.AreEqual(null, ArcPaths.GetCleanedDirectoryPath(null));
+            Assert.AreEqual(null, ArcPaths.GetCleanedFilePath(null));
         }
 
         [TestMethod]
         public void EmptyString()
         {
-            Assert.AreEqual("", ArcPaths.GetCleanedDirectoryPath(""));
+            Assert.AreEqual("", ArcPaths.GetCleanedFilePath(""));
         }
 
         [TestMethod]
         public void RootDirectory()
         {
-            Assert.AreEqual("item", ArcPaths.GetCleanedDirectoryPath("item"));
+            Assert.AreEqual("item", ArcPaths.GetCleanedFilePath("item"));
         }
 
         [TestMethod]
         public void TrailingSlash()
         {
-            Assert.AreEqual("b/", ArcPaths.GetCleanedDirectoryPath("b/"));
+            Assert.AreEqual("b/", ArcPaths.GetCleanedFilePath("b/"));
         }
 
         [TestMethod]
         public void DoubleForwardSlash()
         {
-            Assert.AreEqual("a/b/", ArcPaths.GetCleanedDirectoryPath("a//b/"));
+            Assert.AreEqual("a/b/", ArcPaths.GetCleanedFilePath("a//b/"));
         }
 
         [TestMethod]
         public void DoubleBackwardSlash()
         {
-            Assert.AreEqual("a/b/", ArcPaths.GetCleanedDirectoryPath("a\\b/"));
+            Assert.AreEqual("a/b/", ArcPaths.GetCleanedFilePath("a\\b/"));
         }
 
         [TestMethod]
         public void DoubleForwardAndBackwardSlashes()
         {
-            Assert.AreEqual("a/b/c/d/", ArcPaths.GetCleanedDirectoryPath("a\\b//c\\d//"));
+            Assert.AreEqual("a/b/c/d/", ArcPaths.GetCleanedFilePath("a\\b//c\\d//"));
         }
     }
 }